% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_table.R
\name{add_anbr}
\alias{add_anbr}
\title{add_anbr}
\usage{
add_anbr(df, env = parent.frame())
}
\arguments{
\item{df}{dataframe}

\item{env}{environment}
}
\value{
df with rowby row header added
}
\description{
add_anbr
}
