% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_table.R
\name{add_rowtext_by}
\alias{add_rowtext_by}
\title{add_rowtext_by}
\usage{
add_rowtext_by(df, tablebyvar, env)
}
\arguments{
\item{df}{dataframe}

\item{tablebyvar}{df field that breaks apart table}

\item{env}{environment}
}
\value{
df with rowtext row header added
}
\description{
add_rowtext_by
}
