% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bottom_borders.R
\name{no_borders}
\alias{no_borders}
\alias{spanning_borders}
\alias{col_borders}
\alias{single_border}
\alias{row_border}
\title{Removes all borders from the table}
\usage{
no_borders(ht, matrix = NULL)

spanning_borders(row)

col_borders(col, rows)

single_border(row, col)

row_border(row)
}
\arguments{
\item{row}{\code{numeric} the row of the table}

\item{col}{\code{numeric} the column of the cell}

\item{rows}{\code{numeric} the range of rows to include}
}
\description{
Adds borders under cells that are not empty.
The borders do not touch each other - they are separate.
}
