% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdisc advs}
\alias{cdisc advs}
\alias{cdisc_advs}
\title{\code{ADVS} data created from subsetting the \code{CDISC} \code{ADVS} dataset}
\format{
A data frame with 1938 rows and 35 variables:
\describe{
\item{\code{STUDYID}}{Study Identifier}
\item{\code{SITEID}}{Study Site Identifier}
\item{\code{USUBJID}}{Unique Subject Identifier}
\item{\code{AGE}}{Age}
\item{\code{AGEGR1}}{Pooled Age Group 1}
\item{\code{AGEGR1N}}{Pooled Age Group 1 (N)}
\item{\code{RACE}}{Race}
\item{\code{RACEN}}{Race (N)}
\item{\code{SEX}}{Sex}
\item{\code{SAFFL}}{Safety Population Flag}
\item{\code{TRTSDT}}{Date of First Exposure to Treatment}
\item{\code{TRTEDT}}{Date of Last Exposure to Treatment}
\item{\code{TRTP}}{Planned  Treatment}
\item{\code{TRTPN}}{Planned  Treatment (N)}
\item{\code{TRTA}}{Actual Treatment}
\item{\code{TRTAN}}{Actual Treatment (N)}
\item{\code{PARAMCD}}{Parameter Code}
\item{\code{PARAM}}{Parameter}
\item{\code{PARAMN}}{Parameter (N)}
\item{\code{ADT}}{Analysis Date}
\item{\code{ADY}}{Analysis Relative Day}
\item{\code{ATPTN}}{Analysis Timepoint (N)}
\item{\code{ATPT}}{Analysis Timepoint}
\item{\code{AVISIT}}{Analysis Visit}
\item{\code{AVISITN}}{Analysis Visit (N)}
\item{\code{AVAL}}{Analysis Value}
\item{\code{BASE}}{Baseline Value}
\item{\code{BASETYPE}}{Baseline Value}
\item{\code{CHG}}{Change from Baseline}
\item{\code{PCHG}}{Percent Change from Baseline}
\item{\code{VISITNUM}}{Visit Number}
\item{\code{VISIT}}{Visit Name}
\item{\code{VSSEQ}}{Sequence Number}
\item{\code{ANL01FL}}{Analysis 01 - Special Interest Flag}
\item{\code{ABLFL}}{Baseline Record Flag}
}
}
\source{
CDISC SDTM/ADAM Pilot Project.
}
\usage{
cdisc_advs
}
\description{
\code{ADVS} data created from subsetting the \code{CDISC} \code{ADVS} dataset
}
\keyword{CDISC}
\keyword{advs}
\keyword{datasets}
