% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_alpha}
\alias{check_alpha}
\title{check_alpha}
\usage{
check_alpha(alpha)
}
\arguments{
\item{alpha}{argument for univar}
}
\value{
NULL
Throws error if alpha is not between 0 and 1
}
\description{
check_alpha
}
