% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_column_metadata}
\alias{check_column_metadata}
\title{check_table_metadata}
\usage{
check_column_metadata(column_metadata)
}
\arguments{
\item{column_metadata}{to check for format issues}
}
\value{
NULL
Throws error if any column_metadata has a missing required field
}
\description{
check_table_metadata
}
