% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_column_metadata_file}
\alias{check_column_metadata_file}
\title{check_column_metadata_file}
\usage{
check_column_metadata_file(column_metadata_file)
}
\arguments{
\item{column_metadata_file}{meta data file to check}
}
\value{
NULL
Throws error if file does not exist then throws another error if file exists but has incorrect version
}
\description{
check_column_metadata_file
}
