% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_dfs}
\alias{check_dfs}
\title{check_dfs}
\usage{
check_dfs(dfs)
}
\arguments{
\item{dfs}{to check if provided}
}
\value{
NULL
Throws error if any df has a missing required field
}
\description{
check_dfs
}
