% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_dfs_formats}
\alias{check_dfs_formats}
\title{check_dfs_formats}
\usage{
check_dfs_formats(dfs)
}
\arguments{
\item{dfs}{to check for format issues}
}
\value{
NULL
Throws error if any df has a missing required field
}
\description{
check_dfs_formats
}
