% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_file_exists}
\alias{check_file_exists}
\title{check_file_exists}
\usage{
check_file_exists(func, arglist, file)
}
\arguments{
\item{func}{argument base function name}

\item{arglist}{required argument to check}

\item{file}{required arguments for function}
}
\value{
NULL
Throws error file missing
}
\description{
check_file_exists
}
