% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_freq_cutoff}
\alias{check_freq_cutoff}
\title{check_freq_cutoff}
\usage{
check_freq_cutoff(df)
}
\arguments{
\item{df}{freq dataframe}
}
\value{
NULL
Throws error if cutoff results in empty df, called by freq
}
\description{
check_freq_cutoff
}
