% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_geo_stats}
\alias{check_geo_stats}
\title{check_geo_stats}
\usage{
check_geo_stats(arglist)
}
\arguments{
\item{arglist}{arguments to check}
}
\value{
NULL
Throws error if geo stats can not be calculated and are present
}
\description{
check_geo_stats
}
