% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_in_df}
\alias{check_in_df}
\title{check_in_df}
\usage{
check_in_df(
  func,
  arglist,
  df_field = "df",
  args_that_require_check = c("colvar", "rowvar", "rowbyvar", "tablebyvar",
    "precisionby", "precisionon", "groupby", "tableby", "idvars", "var")
)
}
\arguments{
\item{func}{argument base function name}

\item{arglist}{argument to check}

\item{df_field}{which dataframe to test}

\item{args_that_require_check}{required argument to check}
}
\value{
NULL
Throws error if arg does not exist in df
}
\description{
check_in_df
}
