% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_in_df_bind_table}
\alias{check_in_df_bind_table}
\title{check_in_df_bind_table}
\usage{
check_in_df_bind_table(func, dfs, arglist)
}
\arguments{
\item{func}{argument base function name}

\item{dfs}{list of dfs to check}

\item{arglist}{required argument to check}
}
\value{
NULL
Throws error if arg does not exist in df
}
\description{
check_in_df_bind_table
}
