% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_nested_rowvar}
\alias{check_nested_rowvar}
\title{check_nested_rowvar}
\usage{
check_nested_rowvar(func, arglist, df_field = "df")
}
\arguments{
\item{func}{argument base function name}

\item{arglist}{required argument to check}

\item{df_field}{required df to check}
}
\value{
NULL
Throws error if arg does not exist in df
}
\description{
check_nested_rowvar
}
