% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_nested_vars}
\alias{check_nested_vars}
\title{check_nested_vars}
\usage{
check_nested_vars(nested_vars)
}
\arguments{
\item{nested_vars}{argument that has nested vars to check}
}
\value{
TRUE/FALSE if argument is structured correctly
}
\description{
check_nested_vars
}
