% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_opath}
\alias{check_opath}
\title{check_opath}
\usage{
check_opath(opath)
}
\arguments{
\item{opath}{argument to check if they exist}
}
\value{
NULL
Throws error if opath dies not exist
}
\description{
check_opath
}
