% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_orientation}
\alias{check_orientation}
\title{check_orientation}
\usage{
check_orientation(orientation)
}
\arguments{
\item{orientation}{argument to check if they exist}
}
\value{
NULL
Throws error if orientation is incorrect
}
\description{
check_orientation
}
