% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_plotnames}
\alias{check_plotnames}
\title{check_plotnames}
\usage{
check_plotnames(plotnames)
}
\arguments{
\item{plotnames}{plotname to check}
}
\value{
NULL
Throws error if \code{plotnames} do not exist
}
\description{
check_plotnames
}
