% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_req_fields}
\alias{check_req_fields}
\title{check_req_fields}
\usage{
check_req_fields(func, arglist, req_fields)

check_req_fields(func, arglist, req_fields)
}
\arguments{
\item{func}{argument base function name}

\item{arglist}{required argument to check}

\item{req_fields}{fields in df that are required}
}
\value{
NULL
Throws error if arg does not exist in df

NULL
Throws error if arg does not exist in df
}
\description{
check_req_fields

check_req_fields
}
