% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_req_vars}
\alias{check_req_vars}
\title{check_req_vars}
\usage{
check_req_vars(func, arglist, lst_req)
}
\arguments{
\item{func}{argument base function name}

\item{arglist}{required argument to check}

\item{lst_req}{required arguments for function}
}
\value{
NULL
Throws error if required arg is missing
}
\description{
check_req_vars
}
