% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_statlist}
\alias{check_statlist}
\title{check_statlist}
\usage{
check_statlist(func, astatlist)
}
\arguments{
\item{func}{argument base function name}

\item{astatlist}{statlist to test}
}
\value{
NULL
Throws error if statlist contains incorrect stat or stats
}
\description{
check_statlist
}
