% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_statlist_N}
\alias{check_statlist_N}
\title{check_statlist_N}
\usage{
check_statlist_N(func, astatlist)
}
\arguments{
\item{func}{argument base function name}

\item{astatlist}{statlist to test}
}
\value{
NULL
Throws error if N is used with no other args used by freq functions
}
\description{
check_statlist_N
}
