% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_subset}
\alias{check_subset}
\title{check_subset}
\usage{
check_subset(func, subset, df)
}
\arguments{
\item{func}{argument base function name}

\item{subset}{subset to check}

\item{df}{dataframe that will be filtered}
}
\value{
NULL
Throws error if alpha is not between 0 and 1
}
\description{
check_subset
}
