% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_table_metadata}
\alias{check_table_metadata}
\title{check_table_metadata}
\usage{
check_table_metadata(table_metadata, env)
}
\arguments{
\item{table_metadata}{to check for format issues}

\item{env}{to check for values}
}
\value{
NULL
Throws error if any column_metadata has a missing required field
}
\description{
check_table_metadata
}
