% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_var_duplicates}
\alias{check_var_duplicates}
\title{check_var_duplicates}
\usage{
check_var_duplicates(func, arglist)
}
\arguments{
\item{func}{argument base function name}

\item{arglist}{required argument to check}
}
\value{
NULL
Throws error if arg reference the same value
}
\description{
check_var_duplicates
}
