% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_var_types}
\alias{check_var_types}
\title{check_var_types}
\usage{
check_var_types(func, arglist, lst_types)
}
\arguments{
\item{func}{argument base function name}

\item{arglist}{required argument to check}

\item{lst_types}{required arguments types}
}
\value{
NULL
Throws error if arg is incorrect types
}
\description{
check_var_types
}
