% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_wcol}
\alias{check_wcol}
\title{check_wcol}
\usage{
check_wcol(huxme, wcol)
}
\arguments{
\item{huxme}{dataframe to check}

\item{wcol}{argument to check if they exist}
}
\value{
NULL
Throws error if opath dies not exist
}
\description{
check_wcol
}
