% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bottom_borders.R
\name{col_borders}
\alias{col_borders}
\title{Adds borders under cells in a column}
\usage{
col_borders(col, rows)
}
\arguments{
\item{col}{\code{numeric} the column of the table}

\item{rows}{\code{numeric} the range of rows to include}
}
\description{
Adds borders under cells that are not empty in a column
}
