% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{column_metadata}
\alias{column_metadata}
\title{Metadata describing table column layouts}
\format{
A data frame with one row per column for each table type
and six variables:
\describe{
\item{\code{tbltype}}{identifier used to group a table column layout}
\item{\code{coldef}}{distinct variable values used, typically numeric
and typically a treatment or main effect variable, think \code{TRT01PN}}
\item{\code{decode}}{decode of \code{coldef} that will display as a column header
in the table}
\item{\code{span1}}{spanning header to display across multiple columns}
\item{\code{span2}}{spanning header to display across multiple columns,
second level}
\item{\code{span3}}{spanning header to display across multiple columns,
third level}
}
}
\usage{
column_metadata
}
\description{
This is used by \code{tlgsetup} to prepare you input data to support
the desired column layout.
}
\keyword{datasets}
