% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{tidytlg_titles}
\alias{tidytlg_titles}
\alias{tidytlg_metadata}
\title{Helper functions for returning files used in gentlg}
\usage{
tidytlg_titles(path)

tidytlg_metadata(path)
}
\arguments{
\item{path}{Working directory of the project}
}
\value{
A character vector to the requested file.
}
\description{
Helper functions for returning files used in gentlg
}
