% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileName.R
\name{get_file_name}
\alias{get_file_name}
\title{getFileName}
\usage{
get_file_name()
}
\value{
The file path up until the program's name.
}
\description{
This function returns the file path up until the program's
name.
}
\examples{
get_file_name()
}
\author{
Steven haesendonckx {shaesen2@its.jnj.com}
}
