% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bottom_borders.R
\name{no_borders}
\alias{no_borders}
\title{Removes all borders from the table}
\usage{
no_borders(ht, matrix = NULL)
}
\arguments{
\item{ht}{\code{huxtable} object.}

\item{matrix}{\code{matrix} of bottom borders. Ignored. Included for the sake
of compatibility with the interface of all border mutating functions.}
}
\description{
Removes all borders from the table
}
\seealso{
Other border_functions: 
\code{\link{col_borders}()},
\code{\link{row_border}()},
\code{\link{single_border}()},
\code{\link{spanning_borders}()}
}
\concept{border_functions}
