% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tidytlg.freq}
\alias{print.tidytlg.freq}
\title{Print tidytlg table to the R Console}
\usage{
\method{print}{tidytlg.freq}(x, ...)
}
\description{
When inputting a variable to the console, print is implicitly called on the object.
This S3 method prints out a nice format for the user.
}
