% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_leading_whitespaces_with_indentation.R
\name{replace_leading_whitespaces_with_indentation}
\alias{replace_leading_whitespaces_with_indentation}
\title{Replace leading whitespaces with left indentation RTF markup}
\usage{
replace_leading_whitespaces_with_indentation(x)
}
\arguments{
\item{x}{\code{character(1)} a string to replace the leading whitespaces.}
}
\value{
\code{character(1)} RTF markup with leading whitespaces replaced.
}
\description{
Replace leading whitespaces with left indentation RTF markup
}
\details{
The following function receives a string 'x' and returns a modified string
where every leading whitespace is replaced with 90 twips (0.0625 inches) left indentation RTF markup.
If the input does not start with a whitespace, the string is returned as is.
}
\examples{
replace_leading_whitespaces_with_indentation("    this is x")
# [1] "\\intbl\\li360\\fi0 this is x"
replace_leading_whitespaces_with_indentation("this is x")
# [1] "this is x"

}
\keyword{internal}
