% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{replace_na_with_blank}
\alias{replace_na_with_blank}
\title{Replace NA with ""}
\usage{
replace_na_with_blank(x)
}
\arguments{
\item{x}{variable to check for NA and replace with "".}
}
\value{
x with NA's replaced with "".  Factors will add ""
as the first level.
}
\description{
Used to swap in "" for by variables so the headers sort correctly to the top
}
\examples{
replace_na_with_blank(c("a", "b", NA))

replace_na_with_blank(factor(c("a", "b", NA), levels = c("a", "b")))
}
