% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{roxygen2_data}
\alias{roxygen2_data}
\title{roxygen2_data}
\usage{
roxygen2_data(df, study = NULL, pkg = "~/tidytlg")
}
\arguments{
\item{df}{data frame.}

\item{study}{study name.}

\item{pkg}{package path.}
}
\value{
Roxygen2 documentation for the data frame.
}
\description{
Creates the Roxygen2 documentation for a given data frame.
}
\examples{
\dontrun{
  ds <- list.files(path = "~/reporter/data")
  ds
  roxygen2_data(cdisc_adsl)
  roxygen2_data(cdisc_adae)
  roxygen2_data(cdisc_advs)
  roxygen2_data(cdisc_adlbc)
  roxygen2_data(cdisc_adlbh)
  roxygen2_data(cdisc_adlbhy)
  roxygen2_data(cdisc_adqsadas)
  roxygen2_data(cdisc_adqscibc)
  roxygen2_data(cdisc_adqsnpix)
  roxygen2_data(cdisc_adtte)
 }
}
\author{
Steven Haesendonckx {shaesen2@its.jnj.com}
}
