% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bottom_borders.R
\name{single_border}
\alias{single_border}
\title{Adds a border under a cell}
\usage{
single_border(row, col)
}
\arguments{
\item{row}{\code{numeric} the row of the cell}

\item{col}{\code{numeric} the column of the cell}
}
\description{
Adds a border under a cell
}
