% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bottom_borders.R
\name{spanning_borders}
\alias{spanning_borders}
\title{Adds borders under cells in a row}
\usage{
spanning_borders(row)
}
\arguments{
\item{row}{\code{numeric} the row of the table}
}
\description{
Adds borders under cells that are not empty.
The borders do not touch each other - they are separate.
}
