% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar.R
\name{univar}
\alias{univar}
\title{Descriptive statistics}
\usage{
univar(
  df,
  colvar = NULL,
  tablebyvar = NULL,
  rowvar = NULL,
  rowbyvar = NULL,
  statlist = getOption("tidytlg.univar.statlist.default"),
  decimal = 1,
  precisionby = NULL,
  precisionon = NULL,
  wide = FALSE,
  alpha = 0.05,
  rowtext = NULL,
  row_header = NULL,
  .keep = TRUE,
  .ord = FALSE,
  ...
)
}
\arguments{
\item{df}{(required) dataframe containing records to summarize by treatment}

\item{colvar}{(required) character vector of the treatment variable within
the dataframe}

\item{tablebyvar}{(optional) repeat entire table by variable within df}

\item{rowvar}{(required) character vector of variable to summarize within the
dataframe}

\item{rowbyvar}{(optional) repeat \code{rowvar} by variable within df}

\item{statlist}{(optional) statlist object of stats to keep (default =
statlist(c("N", "MEANSD", "MEDIAN", "RANGE", "IQRANGE")))}

\item{decimal}{(optional) decimal precision root level, when using
\code{presisionby} this will be used as the base decimal cap (default = 1)}

\item{precisionby}{(optional) vector of by variable(s) to use when
calculating parameter based precision}

\item{precisionon}{(optional) variable to use when calculating parameter
based precision. If \code{precisionby} is specified but not \code{precisionon} this will
default to \code{rowvar}}

\item{wide}{(optional) logical indicating to convert labels to column and
columns to labels (default = FALSE)}

\item{alpha}{(optional) alpha level for 2-sided confidence interval (default
= 0.05)}

\item{rowtext}{(optional) A text string to replace the \code{label} value on the
table. Useful for tables with a single row.}

\item{row_header}{(optional) A row to add as a header for the table.}

\item{.keep}{(optional) Should the \code{rowbyvar} and \code{tablebyvar} be output in the
table.  If FALSE, \code{rowbyvar} will still be output in the \code{label} column.
(default = TRUE)}

\item{.ord}{Should the ordering columns be output with the table? This is
useful if a table needs to be merged or reordered in any way after build.}

\item{...}{(optional) Named arguments to be included as columns on the table.}
}
\value{
dataframe of results
}
\description{
Univariate statitstics for a variables by treatment and/or group.
}
