% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdisc adae}
\alias{cdisc adae}
\alias{cdisc_adae}
\title{ADAE data created from subsetting the CDISC ADAE dataset}
\format{
A data frame with 84 rows and 55 variables:
\describe{
\item{STUDYID}{Study Identifier}
\item{SITEID}{Study Site Identifier}
\item{USUBJID}{Unique Subject Identifier}
\item{SUBJID}{Subject Identifier for the Study}
\item{TRTA}{Actual Treatment}
\item{TRTAN}{Actual Treatment (N)}
\item{AGE}{Age}
\item{AGEGR1}{Pooled Age Group 1}
\item{AGEGR1N}{Pooled Age Group 1 (N)}
\item{RACE}{Race}
\item{RACEN}{Race (N)}
\item{SEX}{Sex}
\item{SAFFL}{Safety Population Flag}
\item{TRTSDT}{Date of First Exposure to Treatment}
\item{TRTEDT}{Date of Last Exposure to Treatment}
\item{ASTDT}{Analysis Start Date}
\item{ASTDTF}{Analysis Start Date Imputation Flag}
\item{ASTDY}{Analysis Start Relative Day}
\item{AENDT}{Analysis End Date}
\item{AENDY}{Analysis End Relative Day}
\item{ADURN}{AE Duration (N)}
\item{ADURU}{AE Duration Units}
\item{AETERM}{Reported Term for the Adverse Event}
\item{AELLT}{Lowest Level Term}
\item{AELLTCD}{Lowest Level Term Code}
\item{AEDECOD}{Dictionary-Derived Term}
\item{AEPTCD}{Preferred Term Code}
\item{AEHLT}{High Level Term}
\item{AEHLTCD}{High Level Term Code}
\item{AEHLGT}{High Level Group Term}
\item{AEHLGTCD}{High Level Group Term Code}
\item{AEBODSYS}{Body System or Organ Class}
\item{AESOC}{Primary System Organ Class}
\item{AESOCCD}{Primary System Organ Class Code}
\item{AESEV}{Severity/Intensity}
\item{AESER}{Serious Event}
\item{AESCAN}{Involves Cancer}
\item{AESCONG}{Congenital Anomaly or Birth Defect}
\item{AESDISAB}{Persist or Signif Disability/Incapacity}
\item{AESDTH}{Results in Death}
\item{AESHOSP}{Requires or Prolongs Hospitalization}
\item{AESLIFE}{Is Life Threatening}
\item{AESOD}{Occurred with Overdose}
\item{AEREL}{Causality}
\item{AEACN}{Action Taken with Study Treatment}
\item{AEOUT}{Outcome of Adverse Event}
\item{AESEQ}{Sequence Number}
\item{TRTEMFL}{Treatment Emergent Analysis Flag}
\item{AOCCFL}{1st Occurrence of Any AE Flag}
\item{AOCCSFL}{1st Occurrence of SOC Flag}
\item{AOCCPFL}{1st Occurrence of Preferred Term Flag}
\item{AOCC02FL}{1st Occurrence 02 Flag for Serious}
\item{AOCC03FL}{1st Occurrence 03 Flag for Serious SOC}
\item{AOCC04FL}{1st Occurrence 04 Flag for Serious PT}
\item{CQ01NAM}{Customized Query 01 Name}
\item{AOCC01FL}{1st Occurrence 01 Flag for CQ01}
}
}
\source{
CDISC SDTM/ADAM Pilot Project.
}
\usage{
cdisc_adae
}
\description{
ADAE data created from subsetting the CDISC ADAE dataset
}
\keyword{CDISC}
\keyword{adae}
\keyword{datasets}
