% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bottom_borders.R
\name{col_borders}
\alias{col_borders}
\title{Adds borders under cells in a column}
\usage{
col_borders(col, rows)
}
\arguments{
\item{col}{\code{numeric} the column of the table}

\item{rows}{\code{numeric} the range of rows to include}
}
\description{
Adds borders under cells in a column
}
\seealso{
Other border_functions: 
\code{\link{no_borders}()},
\code{\link{row_border}()},
\code{\link{single_border}()},
\code{\link{spanning_borders}()}
}
\concept{border_functions}
