% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bottom_borders.R
\name{old_format}
\alias{old_format}
\title{Adds bottom borders according to the old formatting}
\usage{
old_format(ht, colspan, colheader, tlf)
}
\arguments{
\item{ht}{the hux object passed to \code{\link[=gentlg]{gentlg()}}}

\item{colspan}{\code{colspan} argument to \code{\link[=gentlg]{gentlg()}}}

\item{colheader}{\code{colheader} argument to \code{\link[=gentlg]{gentlg()}}}

\item{tlf}{\code{character} type of the output}
}
\value{
a bottom border matrix for use with \code{\link[=add_bottom_borders]{add_bottom_borders()}} or \code{NULL}
if \code{ht} is \code{NULL}
}
\description{
This function is vectorized over its arguments.
}
\keyword{internal}
