% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bottom_borders.R
\name{row_border}
\alias{row_border}
\title{Adds a continuous bottom border under a row}
\usage{
row_border(row)
}
\arguments{
\item{row}{\code{numeric} the row of the table}
}
\description{
Adds a continuous bottom border under a row
}
\seealso{
Other border_functions: 
\code{\link{col_borders}()},
\code{\link{no_borders}()},
\code{\link{single_border}()},
\code{\link{spanning_borders}()}
}
\concept{border_functions}
