% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bottom_borders.R
\name{spanning_borders}
\alias{spanning_borders}
\title{Adds borders under cells in a row, excluding the first column.}
\usage{
spanning_borders(row, cols = c(-1))
}
\arguments{
\item{row}{\code{numeric} the row of the table}

\item{cols}{\code{numeric} the columns of the row to consider}
}
\description{
Adds borders under cells that are not empty in a given row,
omitting the first column of the row.
The borders do not touch each other - they are separate.
}
\seealso{
Other border_functions: 
\code{\link{col_borders}()},
\code{\link{no_borders}()},
\code{\link{row_border}()},
\code{\link{single_border}()}
}
\concept{border_functions}
